Filter functionality Done

+++ src/test/java/teamcalypso/FilterTest.java
@@ -42,7 +42,7 @@ public class FilterTest {
         Filter firstFilter = new Filter(data);
         Storage[] array = firstFilter.applyFilter();
         assertEquals(array.length,1);
-        assertEquals(array[0].getType(),"int");
+        assertEquals(array[0].getType(),"double");
     }
 
     @Test
